refactor nav into nav folder
