+++ components/nav/NavBar.tsx
@@ -4,7 +4,7 @@ import useSWR from 'swr';
 import { useSidebarContext } from '~/lib/context/sidebar_context';
 import { SiteList } from '~/lib/types';
 
-import { MenuLogo, NowcastingLogo } from './icons/navbar_icons';
+import { MenuLogo, NowcastingLogo } from '../icons/navbar_icons';
 
 const NavBar: FC = () => {
   const { isSidebarOpen, openSidebar } = useSidebarContext();
