+++ components/nav/index.tsx
@@ -1,0 +1,5 @@
+import BottomNavBar from "./BottomNavBar";
+import NavBar from "./NavBar";
+import Sidebar from "./SideBar";
+
+export {BottomNavBar, NavBar, Sidebar}
